+++ src/types/StoryMeta.d.ts
@@ -1,3 +1,4 @@
 export type StoryMeta = {
   name: string
+  documentation?: string
 }