+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added ability to set a documentation for stories
+  - This can be set with the `documentation` property in `defineStory`
+
 ## [0.0.2] - 2025-01-11
 ### Fixed
 - Fixed intergation with [Nuxt i18n](https://i18n.nuxtjs.org/)